refactor run-time detection module
