+++ crates/stdsimd/src/lib.rs
@@ -28,3 +28,10 @@ extern crate std;
 mod stdsimd;
 
 pub use stdsimd::*;
+
+#[allow(unused_imports)]
+use _std::prelude;
+#[allow(unused_imports)]
+use _std::fs;
+#[allow(unused_imports)]
+use _std::io;
