+++ stdsimd/arch/detect/os/linux/mips.rs
@@ -1,0 +1,36 @@
+//! Run-time feature detection for MIPS on Linux.
+
+#[path = "../cache.rs"]
+mod cache;
+#[path = "../bit.rs"]
+mod bit;
+#[path = "auxvec.rs"]
+mod auxvec;
+
+use arch::detect::Feature;
+
+/// Performs run-time feature detection.
+pub fn check_for(x: Feature) -> bool {
+    cache::test(x as u32, detect_features)
+}
+
+/// Try to read the features from the auxiliary vector, and if that fails, try
+/// to read them from /proc/cpuinfo.
+fn detect_features() -> cache::Initializer {
+    let mut value = cache::Initializer::default();
+    let enable_feature = |value: &mut cache::Initializer, f, enable| {
+        if enable {
+            value.set(f as u32);
+        }
+    };
+
+    // The values are part of the platform-specific [asm/hwcap.h][hwcap]
+    //
+    // [hwcap]: https://github.com/torvalds/linux/blob/master/arch/arm64/include/uapi/asm/hwcap.h
+    if let Ok(auxv) = auxvec::auxv() {
+        enable_feature(&mut value, Feature::msa, bit::test(auxv.hwcap, 1));
+        return value;
+    }
+    // TODO: fall back via cpuinfo
+    value
+}
