+++ stdsimd/arch/detect/os/other.rs
@@ -1,0 +1,8 @@
+//! Other operating systems
+
+use arch::detect::Feature;
+
+/// Performs run-time feature detection.
+pub fn check_for(_x: Feature) -> bool {
+    false
+}
